services/websocket/pyproject.toml
[project]
name = "websocket-service"
version = "1.0.0"
description = "WebSocket real-time sync service for Phase V cloud-native architecture"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "dapr>=1.12.0",
    "dapr-ext-fastapi>=1.12.0",
    "pydantic>=2.9.0",
    "pyjwt[crypto]>=2.9.0",  # JWT validation for WebSocket auth
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
